+++ media/media-viewer/src/viewers/useThemeObserverHoc.tsx
@@ -3,7 +3,7 @@ import { useThemeObserver } from '@atlaskit/tokens';
 
 const withThemeObserverHOC = (Component: any) => {
   return (props: any) => {
-    const theme = useThemeObserver();
+    const { colorMode: theme } = useThemeObserver();
 
     return <Component theme={theme} {...props} />;
   };